FT-17(update) split the files
