+++ src/js/templates/render-backdrop.js
@@ -1,0 +1,3 @@
+export default function renderBackdrop() {
+  return `<div class="backdrop is-hidden"></div>`;
+}
